Fixed delete tests
